+++ src/wai/annotations/isp/rename/component/__init__.py
@@ -1,0 +1,1 @@
+from ._Rename import Rename