+++ src/wai/annotations/isp/rename/specifier/__init__.py
@@ -1,0 +1,1 @@
+from ._RenameISPSpecifier import RenameISPSpecifier